+++ index.html
@@ -5,8 +5,8 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+  <script type="module" src="js/main.js"></script>
   <title>Кексобукинг</title>
-  <script src="js/main.js"></script>
 </head>
 <body>
